feat:Side navigation treeview in note application - MEED-10100 - Meeds-io/MIPs#235
This change will implement a new side bar navigation for note application
