+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_en.properties
@@ -175,6 +175,11 @@ notes.label.no-content.no-redactor.content.last=Or create a new note
 notes.label.noteNotFound=We could not find this note.
 notes.label.noteSearchNotFound=No notes found for:
 notes.label.noteNotFound.button=Back to Notes App
+
+notes.label.tree=Notes tree
+notes.tooltip.close.tree=Close notes tree
+notes.tooltip.open.tree=Open notes tree
+
 notes.message.missingImportNotes=Import Notes is mandatory.
 import.drawer.maxFileSize.error=You can’t upload files larger than {0} MB
 import.drawer.fileType.error=You should upload a zip note
