+++ notes-webapp/src/main/webapp/vue-app/notes-treeview/initComponents.js
@@ -1,8 +1,10 @@
 import NoteBreadcrumb from '../notes/components/NoteBreadcrumb.vue';
 import NoteTreeviewDrawer from './components/NoteTreeviewDrawer.vue';
+import NoteTreeViewSideBar from './components/NoteTreeViewSideBar.vue';
 
 const components = {
   'note-treeview-drawer': NoteTreeviewDrawer,
+  'note-treeview-sideBar': NoteTreeViewSideBar,
   'note-breadcrumb': NoteBreadcrumb
 };
